server/requirements.txt
# Dummy requirements.txt
PyYAML
